removed some simple warnings. two types of warnings still remain: 'int to char' and 'int to unsigned int'

+++ SYMPHONY/include/sym_timemeas.h
@@ -96,7 +96,7 @@
 
 #define DBLTOTV(x, tv)						\
      tv.tv_sec = (int) floor(x);				\
-     tv.tv_usec = (int) floor(1000000 * (x - tv.tv_sec));
+     tv.tv_usec = (int) floor(1000000 * (x - (double)tv.tv_sec));
 #define DBLTOPTV(x, tvp)						\
      (tvp)->tv_sec = (int) floor(x);					\
      (tvp)->tv_usec = (int) floor(1000000 * (x - (tvp)->tv_sec));
